Banking system(Project)/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Flask-Login==0.5.0
werkzeug==2.0.1
python-dotenv==0.19.0
